moves channel to a separate file
